requirements.txt
langchain
langchain-community
langchain-openai
langchain-google-genai
python-dotenv
streamlit
qdrant-client
google-generativeai
sentence-transformers
unstructured
python-magic
tiktoken
PyPDF2
openpyxl
chromadb
